Move the built-in profile-menu rows into the registry
Notifications, My profile and Log out become providers like any third-party
entry, so a bundle can position itself against them, reorder them, or hide one
via overrideSlotProvider — an SSO-only or kiosk deployment has a reason to drop
Log out.

Logout sits at 9999 rather than a merely large number: the registry defaults a
missing priority to 999, so anything lower loses its last place to a provider
that omits one.

The send-notification modal moves to the modal holder, which is what lets the
notifications row leave UserMenu at all.
