+++ assets/js/src/core/modules/app/base-layout/user-menu/user-menu.tsx
@@ -9,23 +9,12 @@
  */
 
 import { Badge } from '@Pimcore/components/badge/badge'
-import { Dropdown, type DropdownMenuProps, type ItemType } from '@Pimcore/components/dropdown/dropdown'
-import { Icon } from '@Pimcore/components/icon/icon'
-import trackError, { ApiError } from '@Pimcore/modules/app/error-handler'
-import { useLogoutMutation } from '@Pimcore/modules/auth/authorization-api-slice.gen'
+import { Dropdown, type DropdownMenuProps } from '@Pimcore/components/dropdown/dropdown'
 import { isAllowed } from '@Pimcore/modules/auth/permission-helper'
-import { NOTIFICATIONS, NOTIFICATION_SETTINGS } from '@Pimcore/modules/notifications'
-import { IconButton } from '@Pimcore/components/icon-button/icon-button'
-import { Button } from '@Pimcore/components/button/button'
-import { Flex } from '@Pimcore/components/flex/flex'
-import { SendNotificationModal } from '@Pimcore/modules/notifications/send-notification/send-notification-modal'
-import { useWidgetManager } from '@sdk/modules/widget-manager'
-import { theme } from 'antd'
-import React, { useEffect, useState } from 'react'
+import React, { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { useStyle } from './user-menu.styles'
 import { UserPermission } from '@Pimcore/modules/auth/enums/user-permission'
-import { USERPROFILE } from '@Pimcore/modules/auth/profile/profile-container'
 import { useUser } from '@Pimcore/modules/auth/hooks/use-user'
 import { Avatar } from '@Pimcore/components/avatar/avatar'
 import { useUserHelper } from '@Pimcore/modules/auth/hooks/use-user-helper'
@@ -39,12 +28,9 @@ interface IUserMenuProps {
 export const UserMenu = ({ className }: IUserMenuProps): React.JSX.Element => {
   const { t } = useTranslation()
   const { styles } = useStyle()
-  const { token } = theme.useToken()
-  const [logout] = useLogoutMutation()
-  const { openMainWidget } = useWidgetManager()
-  const [sendModal, setSendModal] = useState<boolean>(false)
   const user = useUser()
   const { getUserImageById, updateUserImageInState } = useUserHelper()
+  // Shared with the notifications provider; RTK Query dedupes the two subscriptions into one request.
   const { data } = useNotificationGetUnreadCountQuery(undefined, {
     skip: !isAllowed(UserPermission.Notifications)
   })
@@ -61,20 +47,10 @@ export const UserMenu = ({ className }: IUserMenuProps): React.JSX.Element => {
     }
   }, [])
 
-  const handleLogout = (): void => {
-    const logoutTask = logout()
-
-    logoutTask.then(() => {
-      window.location.reload()
-    }).catch((error: Error) => {
-      trackError(new ApiError(error))
-    })
-  }
-
   const notificationCount = data?.unreadNotificationsCount ?? 0
 
   // Providers must register at module init — useMenuItem() hooks run in slot order.
-  const extensionItems: ItemType[] = useContextMenuSlot(contextMenuConfig.userMenu.name, {})
+  const entries = useContextMenuSlot(contextMenuConfig.userMenu.name, {})
     .map((item) => (item !== null && 'icon' in item && item.icon !== undefined
       ? { ...item, icon: <div className={ 'user-menu__item-icon' }>{item.icon}</div> }
       : item))
@@ -93,118 +69,44 @@ export const UserMenu = ({ className }: IUserMenuProps): React.JSX.Element => {
       ),
       type: 'group'
     },
-    {
-      key: 'notifications',
-      label: t('user-menu.notifications'),
-      icon: <div className={ 'user-menu__item-icon' }>
-        <Badge
-          count={ notificationCount }
-          showZero
-          styles={ {
-            indicator: {
-              // Always shown here, but a zero goes grey rather than wearing the accent colour.
-              background: notificationCount > 0 ? token.colorPrimary : token.colorTextQuaternary,
-              width: 20,
-              height: 20,
-              minWidth: 20,
-              lineHeight: '20px',
-              borderRadius: '50%',
-              fontSize: notificationCount > 99 ? 9 : 10,
-              fontWeight: 'normal',
-              padding: 0,
-              boxShadow: 'none'
-            },
-            root: { marginRight: 0 }
-          } }
-        />
-      </div>,
-      onClick: () => { openMainWidget(NOTIFICATIONS) },
-      hidden: !isAllowed(UserPermission.Notifications),
-      // Both actions open something other than the bell, so each stops the row's own onClick.
-      extra: (
-        <Flex
-          align={ 'center' }
-          gap={ 'mini' }
-        >
-          {isAllowed(UserPermission.SendNotifications) && (
-            <Button
-              className={ 'user-menu__item-extra' }
-              onClick={ (e) => {
-                e.stopPropagation()
-                setSendModal(true)
-              } }
-              size={ 'small' }
-            >{t('user-menu.notification.send')}</Button>
-          )}
-          <IconButton
-            icon={ { value: 'settings' } }
-            onClick={ (e) => {
-              e.stopPropagation()
-              openMainWidget(NOTIFICATION_SETTINGS)
-            } }
-            title={ t('notifications.settings.label') }
-            type={ 'text' }
-          />
-        </Flex>
-      )
-    },
-    {
-      key: 'myprofile',
-      label: t('user-menu.my-profile'),
-      icon: <div className={ 'user-menu__item-icon' }><Icon value={ 'user' } /></div>,
-      onClick: () => { openMainWidget(USERPROFILE) }
-    },
-    ...extensionItems,
-    {
-      key: 'logout',
-      label: t('user-menu.log-out'),
-      icon: <div className={ 'user-menu__item-icon' }><Icon value={ 'log-out' } /></div>,
-      onClick: handleLogout
-    }
+    ...entries
   ]
 
   return (
-    <>
-      <Dropdown
-        className={ className }
-        menu={ { items } }
-        overlayClassName={ [styles.userMenu].join(' ') }
-        overlayStyle={ { minWidth: 275 } }
-        trigger={ ['click'] }
+    <Dropdown
+      className={ className }
+      menu={ { items } }
+      overlayClassName={ [styles.userMenu].join(' ') }
+      overlayStyle={ { minWidth: 275 } }
+      trigger={ ['click'] }
+    >
+      {/* No showZero: permanently on screen, so it must go quiet at nothing to report. */}
+      <Badge
+        count={ notificationCount }
+        data-testid="user-menu-avatar-badge"
+        overflowCount={ 99 }
+        size={ 'small' }
+        styles={ {
+          indicator: {
+            // Fixed circle: the default indicator grows with each digit, which reads as a
+            // stretched pill against a 26px avatar. The font steps down instead.
+            width: 16,
+            height: 16,
+            minWidth: 16,
+            lineHeight: '16px',
+            borderRadius: '50%',
+            fontSize: notificationCount > 9 ? 9 : 10,
+            fontWeight: 'normal',
+            padding: 0
+          }
+        } }
       >
-        {/* No showZero: permanently on screen, so it must go quiet at nothing to report. */}
-        <Badge
-          count={ notificationCount }
-          data-testid="user-menu-avatar-badge"
-          overflowCount={ 99 }
-          size={ 'small' }
-          styles={ {
-            indicator: {
-              // Fixed circle: the default indicator grows with each digit, which reads as a
-              // stretched pill against a 26px avatar. The font steps down instead.
-              width: 16,
-              height: 16,
-              minWidth: 16,
-              lineHeight: '16px',
-              borderRadius: '50%',
-              fontSize: notificationCount > 9 ? 9 : 10,
-              fontWeight: 'normal',
-              padding: 0
-            }
-          } }
-        >
-          <Avatar
-            data-testid="user-menu-avatar"
-            size={ 26 }
-            src={ user?.hasImage && user?.image != null ? user?.image : undefined }
-          />
-        </Badge>
-      </Dropdown>
-
-      <SendNotificationModal
-        onClose={ () => { setSendModal(false) } }
-        open={ sendModal }
-      />
-    </>
+        <Avatar
+          data-testid="user-menu-avatar"
+          size={ 26 }
+          src={ user?.hasImage && user?.image != null ? user?.image : undefined }
+        />
+      </Badge>
+    </Dropdown>
   )
 }
