+++ assets/js/src/core/modules/notifications/index.tsx
@@ -14,43 +14,20 @@ import { serviceIds } from '@Pimcore/app/config/services/service-ids'
 import { moduleSystem } from '@Pimcore/app/module-system/module-system'
 import { NotificationsContainer } from './notifications-container'
 import { NotificationSettingsContainer } from './settings/notification-settings-container'
-import { type WidgetManagerTabConfig } from '../widget-manager/widget-manager-slice'
 import { type BackgroundProcessor } from '../background-processor/services/background-processor'
 import { DemoProcess } from './process/demo-process'
 import { staticWidgetRestorer } from '../widget-manager/services/static-widget-restorer'
 import { type ComponentRegistry } from '../app/component-registry/component-registry'
+import { contextMenuConfig } from '../app/context-menu-registry/context-menu-config'
+import { type ContextMenuRegistryInterface } from '../app/context-menu-registry/context-menu-registry'
+import { notificationsUserMenuItemProvider } from './user-menu-item'
 import { NotificationPopup } from './notification-popup/notification-popup'
-import { UserPermission } from '@Pimcore/modules/auth/enums/user-permission'
 import { type DynamicTypeNotificationChannelRegistry } from './dynamic-types/registry/dynamic-type-notification-channel-registry'
 import { type DynamicTypeAbstractNotificationChannel } from './dynamic-types/definitions/dynamic-type-abstract-notification-channel'
 
-export const NOTIFICATIONS: WidgetManagerTabConfig = {
-  component: 'notifications',
-  name: 'Notifications',
-  id: 'notifications',
-  permission: UserPermission.Notifications,
-  config: {
-    translationKey: 'notifications.label',
-    icon: {
-      type: 'name',
-      value: 'notification-read'
-    }
-  }
-}
+import { NOTIFICATIONS, NOTIFICATION_SETTINGS } from './widget-configs'
 
-export const NOTIFICATION_SETTINGS: WidgetManagerTabConfig = {
-  component: 'notification-settings',
-  name: 'Notification settings',
-  id: 'notification-settings',
-  permission: UserPermission.Notifications,
-  config: {
-    translationKey: 'notifications.settings.label',
-    icon: {
-      type: 'name',
-      value: 'settings'
-    }
-  }
-}
+export { NOTIFICATIONS, NOTIFICATION_SETTINGS } from './widget-configs'
 
 moduleSystem.registerModule({
   onInit: () => {
@@ -74,6 +51,10 @@ moduleSystem.registerModule({
       name: 'notification-popup',
       component: NotificationPopup
     })
+
+    container
+      .get<ContextMenuRegistryInterface>(serviceIds['App/ContextMenuRegistry/ContextMenuRegistry'])
+      .registerToSlot(contextMenuConfig.userMenu.name, notificationsUserMenuItemProvider)
 
     const BackgroundProcessor = container.get<BackgroundProcessor>(serviceIds.backgroundProcessor)
     BackgroundProcessor.registerProcess(new DemoProcess())
